Split xcb_ffi into multiple files
The module structure stays the same way as it was before, but this
commit moves some inner modules to separate files.
